requirements.txt
google-generativeai>=0.5.0
python-dotenv>=1.0.0
python-docx>=1.1.0
reportlab>=4.0.0
matplotlib>=3.8.0
pandas>=2.1.0
numpy>=1.26.0
